# Work_Assignment/AppServer/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "PlugGoBeyond",
    "description" : "Rest API of the mobile app PlugGoBeyond of the project course LI IV in the program Mestrado Integrado em Engenharia Informática",
    "termsOfService" : "http://swagger.io/terms",
    "contact" : {
      "email" : "a34900/a82888/a85731/a86618/a89982@alunos.uminho.pt"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
...
  "servers" : [ {
    "url" : "/"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/vehicleModels" : {
      "get" : {
        "tags" : [ "vehicleModels" ],
        "summary" : "List all vehiclesModels.",
        "description" : "Find all the models of eletric vehicles.",
        "operationId" : "findVehiclesModels",
        "responses" : {
          "200" : {
            "description" : "Success.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/VehicleModels"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Operation error."
          }
        }
      },
      "post" : {
        "tags" : [ "vehicleModels" ],
        "summary" : "Add a vehicle's model into the system.",
        "description" : "Inserts a new model into the system.",
        "operationId" : "addVehicleModel",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VehicleModels"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Success."
          },
          "400" : {
            "description" : "Invalid Parameter."
          }
        }
      }
    },
    "/vehicleModels/{typeCode}" : {
      "get" : {
        "tags" : [ "vehicleModels" ],
        "summary" : "Finds a model of a vehicle with a certain typeCode.",
        "description" : "Returns the model that has a specific typeCode.",
        "operationId" : "findVehicleModel",
        "parameters" : [ {
          "name" : "typeCode",
          "in" : "path",
          "description" : "TypeCode of the model that we want to find.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VehicleModels"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid TypeCode."
          },
          "404" : {
            "description" : "Model of vehicle not found."
          }
        }
      },
      "put" : {
        "tags" : [ "vehicleModels" ],
        "summary" : "Updates a vehicle's model.",
        "operationId" : "updateVehicleModel",
        "parameters" : [ {
          "name" : "typeCode",
          "in" : "path",
          "description" : "TypeCode of the model that needs to be updated.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VehicleModels"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Invalid TypeCode."
          },
          "404" : {
            "description" : "Model of Vehicle not found."
          }
        }
      },
      "delete" : {
        "tags" : [ "vehicleModels" ],
        "summary" : "Removes a vehicle.",
        "description" : "Deletes a model.",
        "operationId" : "deleteVehicleModel",
        "parameters" : [ {
          "name" : "typeCode",
          "in" : "path",
          "description" : "typeCode of the model that needs to be removed.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid typeCode."
          },
          "404" : {
            "description" : "TypeCode not found."
          }
        }
      }
    },
    "/chargingStationWaitingValidation" : {
      "get" : {
        "tags" : [ "stationWithoutValidation" ],
        "summary" : "List all Charging Stations Waiting Validation.",
        "description" : "Returns all stations that were not yet considered valid by the system.",
        "operationId" : "findStationWaitingValidation",
        "responses" : {
          "200" : {
            "description" : "Success.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ChargingStationWaitingValidation"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Operation error."
          }
        }
      },
      "post" : {
        "tags" : [ "stationWithoutValidation" ],
        "summary" : "Add a Stations Waiting Validation into the system.",
        "description" : "Inserts a new station waiting validation.",
        "operationId" : "addStationWithoutValidation",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChargingStationWaitingValidation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Success."
          },
          "400" : {
            "description" : "Invalid Parameter."
          }
        }
      }
    },
    "/chargingStationWaitingValidation/{nameChargingStation}" : {
      "get" : {
        "tags" : [ "stationWithoutValidation" ],
        "summary" : "Finds a station by it's name.",
        "description" : "Given a certain valid name, a station will be returned",
        "operationId" : "findStationWithoutValidationByName",
        "parameters" : [ {
          "name" : "nameChargingStation",
          "in" : "path",
          "description" : "Name of the chargingStation that we want to find.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChargingStationWaitingValidation"
                }
              }
            }
          },
          "400" : {
            "description" : "Operation error."
          },
          "404" : {
            "description" : "Name of station without validation not found."
          }
        }
      },
      "put" : {
        "tags" : [ "stationWithoutValidation" ],
        "summary" : "Updates a Station without validation.",
        "description" : "Given a certain valid name, a station will be updated.",
        "operationId" : "updateStationWithoutValidation",
        "parameters" : [ {
          "name" : "nameChargingStation",
          "in" : "path",
          "description" : "name of the station that needs to be updated.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChargingStationWaitingValidation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Invalid name of the station."
          },
          "404" : {
            "description" : "Name of Station not found."
          }
        }
      },
      "delete" : {
        "tags" : [ "stationWithoutValidation" ],
        "summary" : "Removes a station without validation.",
        "description" : "Deletes a specific station.",
        "operationId" : "deleteStationWithoutValidation",
        "parameters" : [ {
          "name" : "nameChargingStation",
          "in" : "path",
          "description" : "name of the station that needs to be removed.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid name of the station."
          },
          "404" : {
            "description" : "Name of station without validation not found."
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ChargingStationWaitingValidation" : {
        "type" : "object",
        "properties" : {
          "nameChargingStation" : {
            "type" : "string"
          },
          "street" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "locationType" : {
            "type" : "string"
          },
          "accessType" : {
            "type" : "string"
          },
          "restritions" : {
            "type" : "string"
          },
          "aditionalInfo" : {
            "type" : "string"
          }
        },
        "example" : {
          "accessType" : "accessType",
          "nameChargingStation" : "nameChargingStation",
          "city" : "city",
          "street" : "street",
          "locationType" : "locationType",
          "aditionalInfo" : "aditionalInfo",
          "restritions" : "restritions"
        }
      },
...
      "VehicleModels" : {
        "type" : "object",
        "properties" : {
          "typeCode" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "levelTwoCharges" : {
            "type" : "string"
          },
          "fastCharges" : {
            "type" : "string"
          }
        },
        "example" : {
          "name" : "name",
          "fastCharges" : "fastCharges",
          "typeCode" : "typeCode",
          "levelTwoCharges" : "levelTwoCharges"
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}